requirements.txt
Flask==2.0.2
datetime
azure.cosmos